:recycle: refactor(quiz): Apply + Test new source code for Quiz and QuizSet

+++ src/main/java/com/be08/smart_notes/service/NoteService.java
@@ -4,7 +4,14 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.be08.smart_notes.dto.response.NoteResponse;
+import com.be08.smart_notes.exception.AppException;
+import com.be08.smart_notes.exception.ErrorCode;
+import com.be08.smart_notes.mapper.DocumentMapper;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import com.be08.smart_notes.dto.request.NoteUpsertRequest;
@@ -13,18 +20,45 @@ import com.be08.smart_notes.model.Document;
 import com.be08.smart_notes.repository.DocumentRepository;
 
 @Service
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+@Slf4j
 public class NoteService {
-    @Autowired
-    private DocumentRepository documentRepository;
+    AuthorizationService authorizationService;
+    DocumentRepository documentRepository;
+    DocumentMapper documentMapper;
 
-    public Document getNote(int noteId) {
-        Document note = documentRepository.findById(noteId).orElse(null);
-        return note;
+    public NoteResponse getNote(int noteId) {
+        // Get note
+        Document note = documentRepository.findById(noteId).orElseThrow(() -> {
+            log.error("Note with id {} not found", noteId);
+            return new AppException(ErrorCode.DOCUMENT_NOT_FOUND);
+        });
+
+        // Check ownership
+        authorizationService.validateOwnership(note.getUserId());
+
+        return documentMapper.toNoteResponse(note);
     }
 
-    public Document createNote(NoteUpsertRequest newData) {
+    public List<Document> getAllNotesByUserIdAndIds(List<Integer> noteIds) {
+        // Get current user id
+        int currentUserId = authorizationService.getCurrentUserId();
+
+        return documentRepository.findAllByUserIdAndIdIn(currentUserId, noteIds);
+    }
+
+    public List<Document> getAllNotesByIds(List<Integer> noteIds) {
+        return documentRepository.findAllByIdIn(noteIds);
+    }
+
+    public NoteResponse createNote(NoteUpsertRequest newData) {
+        // Get current user id
+        int currentUserId = authorizationService.getCurrentUserId();
+
+        // Create new note
         Document newNote = Document.builder()
-                .userId(newData.getUserId())
+                .userId(currentUserId)
                 .title(newData.getTitle())
                 .content(newData.getContent())
                 .type(DocumentType.NOTE)
@@ -33,23 +67,36 @@ public class NoteService {
                 .build();
 
         documentRepository.save(newNote);
-        return newNote;
+        return documentMapper.toNoteResponse(newNote);
     }
 
-    public void updateNote(int noteId, NoteUpsertRequest updateData) {
-        Document note = documentRepository.findById(noteId).orElse(null);
+    public NoteResponse updateNote(int noteId, NoteUpsertRequest updateData) {
+        // Get note
+        Document note = documentRepository.findById(noteId).orElseThrow(() -> {
+            log.error("Note with id {} not found", noteId);
+            return new AppException(ErrorCode.DOCUMENT_NOT_FOUND);
+        });
+
+        // Check ownership
+        authorizationService.validateOwnership(note.getUserId());
 
         note.setTitle(updateData.getTitle());
         note.setUpdatedAt(LocalDateTime.now());;
         note.setContent(updateData.getContent());
-        documentRepository.save(note);
+        Document updatedNote = documentRepository.save(note);
+        return documentMapper.toNoteResponse(updatedNote);
     }
 
     public void deleteNote(int noteId) {
-        documentRepository.deleteById(noteId);
-    }
+        // Get document
+        Document note = documentRepository.findById(noteId).orElseThrow(() -> {
+            log.error("Note with id {} not found", noteId);
+            throw new AppException(ErrorCode.DOCUMENT_NOT_FOUND);
+        });
 
-    public List<Document> getAllNotesByIds(List<Integer> noteIds) {
-        return documentRepository.findAllByIdIn(noteIds);
+        // Check ownership
+        authorizationService.validateOwnership(note.getUserId());
+
+        documentRepository.deleteById(noteId);
     }
 }
